+++ app/javascript/components/track/approaches-elements/index.ts
@@ -1,0 +1,4 @@
+export { DiggingDeeper } from './DiggingDeeper'
+export { ApproachExamples } from './ApproachExamples'
+export { SectionHeader } from './SectionHeader'
+export { CommunityVideos } from './CommunityVideos'
